+++ appcenter/AppCenter.d.ts
@@ -34,6 +34,8 @@ export function getLogLevel(): Promise<LogLevel | AppCenterLogLevel>;
 export function getSdkVersion(): string;
 export function isEnabled(): Promise<boolean>;
 export function setEnabled(enabled: boolean): Promise<void>;
+export function isNetworkRequestsAllowed(): Promise<boolean>;
+export function setNetworkRequestsAllowed(isAllowed: boolean): Promise<void>;
 export function setUserId(userId: string | null): Promise<void>;
 export function getInstallId(): Promise<string>;
 export function setCustomProperties(properties: CustomProperties): Promise<void>;